15. snmp_cpu_memory_poller/requirements.txt
pysnmp
pyyaml
matplotlib
